Make m_bunchCrossingTool of src/SCTHitEffMonTool.cxx public tool again. (SCT_Monitoring-00-06-98)
	* Make m_bunchCrossingTool of src/SCTHitEffMonTool.cxx public tool again.
	* tag as SCT_Monitoring-00-06-98

2016-12-03 Susumu Oda <Susumu.Oda@cern.ch>
	* Revert TProfile2D_LW to TProfile2D in SCTHitsNoiseMonTool.
	* tag as SCT_Monitoring-00-06-97

2016-11-29 Shunsuke Honda <Shunsuke.Honda@cern.ch>
	* Make NBINS_LBs common variable in the SCT_Monitoring in SCT_MonitoringNumbers
	* Change NBINS_LBs 2000 to 3000
	* tag as SCT_Monitoring-00-06-96

2016-11-28 Shunsuke Honda <Shunsuke.Honda@cern.ch>
	* Change variables TH1D to TH1F and TH? to TH?_LW in the SCTHitsNoiseMonTool(ATLASRECTS-3440)
	* tag as SCT_Monitoring-00-06-95

2016-11-28 Shunsuke Honda <Shunsuke.Honda@cern.ch>
	* Change variables double to float
	* tag as SCT_Monitoring-00-06-94


Former-commit-id: c2aa1c109889ab6f3bcdb667f9be9f6709244061

+++ InnerDetector/InDetMonitoring/SCT_Monitoring/SCT_Monitoring/SCTLorentzMonTool.h
@@ -118,10 +118,10 @@ private:
   //@name Service methods
   //@{
   // Calculate the local angle of incidence
-  int findAnglesToWaferSurface ( const double (&vec)[3], const double &sinAlpha, const Identifier &id, double &theta, double &phi );
+  int findAnglesToWaferSurface ( const float (&vec)[3], const float &sinAlpha, const Identifier &id, float &theta, float &phi );
 
   ///Factory + register for the 2D profiles, returns whether successfully registered
-  Prof_t  pFactory(const std::string & name, const std::string & title, int nbinsx, double xlow, double xhigh, MonGroup & registry, int& iflag);
+  Prof_t  pFactory(const std::string & name, const std::string & title, int nbinsx, float xlow, float xhigh, MonGroup & registry, int& iflag);
   ///Factory + register for the 1D histograms, returns whether successfully registered
   bool h1Factory(const std::string & name, const std::string & title, const float extent, MonGroup & registry, VecH1_t & storageVector);
   //@}
